Add important fix to parameter error.
Numeric hessians from numdifftools are not well behaved at the moment
(for both chi^2 and the fit function).
Tolerances either need to be increased, the second derivative(s) of the
fit function need(s) to be set to 0, or the Hessian for the fit function should be computed analytically.
TODO:
Add an option to enter the first derivatives of the fit function by
hand, GEVP analysis, and double jackknife resampling.
This commit should be nearly at the 1.0 release stage, and should be
considered production ready.