Add feature: save blocks+refactor extraction.
This new feature is needed for fitting under a jackknife
(frozen/double).
Previously, this container was used for EFF_MASS, so extend it to
include not EFF_MASS.
Refactor extraction to accomodate this feature.  Move things out of
unnecessary loops.
This feature costs memory (which is not seriously used at the moment) to
increase speed/elegance.