Add working GEVP. witch GEVP to eff_mass only.
We need to do this because the GEVP is an ill-defined fit:
C(t)v=eigenval*C(t_0)v
Do we fit to some function of t?  t_0 is not a linear function of t, so
this seems problematic.
TODO: use already written EFF_MASS functions to avoid code redundancies.