Allow user to eliminate bad configs from jackknife.
If the user so chooses, they can eliminate jackknife samples.  The
specified eliminated samples are given in a list, e.g. [0,3,4] would
eliminate the first, fourth, and fifth jackknife sample from a given
block.  (In the lattice qcd context, this is useful if the jackknife
blocks were made with bad, perhaps unthermalized (for example),
configs).