Fix remaining aux diagrams;prefer measured over aux
the bug:  aux diagrams had the wrong momenta (needed to multiply
everything by -1 after swapping).  It was fixed for 4 vertices, but now
it's fixed for 2 and 3.

Also, overwrite aux diagrams with measured diagrams (which necessarily
have better statistics.  A goal would be to combine both, if we happen
to have measurements.

Turn off bad fold option (ANTIPERIODIC); It should be eliminated (future commit).