Fix finite box fit function bug:
The bug was that the tsep in the exponentials in the effective mass processor for each dim of the gevp was set to be the tsep of the first dimension.  Thus, if one were to quote the tseps from the config as the fitted tseps, then one would  be in error.  Otherwise, this has no effect on the corretness.  Numerically, the difference is often negligible (and, in fact, beyond machine precision).