Optimize use of pion ratio
for the case of small discretization error vs. statistical error (in this case we can use whatever additive zero we want for each energy level).  Not considered
is the differing excited state contaminations, but small stat. error
should come with the lower states which also have lower excited state
contamination.