Fix bug in allow cut:
The stat. "cut" in this cut is a shortcut to avoid calculating systematic
error if the stat error is already bigger than the best known result.
We should not apply this stat. "cut" to the results we are comparing
against.  (It would be a real cut then, as opposed to a shortcut.)