Add hack to aid in reproduction.
The cut should be applied regardless if it's a legitimate cut, but it
seems likely that the consistency condition is too harsh, so including
points outside the fit window cuts more points.  In other words, the
extra points indicate more should've been cut from the given fit range,
so we have to pretend these points don't exist if we want to reproduce
the given fit range.  Thus, either the cut is illegitimate (too harsh),
or the fit should've actually been thrown out.  Given that p0 has such
problems and the cut causing the problem is the exp. growing cut (which
should not impact p0 at all), we
 should expect that the cut is not legitimate (or else something more
serious is going on).