Relax window equality check (for weak cuts).
The two fit windows should agree since the same fit range should give us a
min for the energy and the corresponding phase shift.  However,
sometimes we observe this not to be the case.  For a sparse number of
results, the fit windows are heavily dependent on a few fit ranges, so
we relax the check to only ask for containment.