Add utility script for copying GEVP elements across the diagonal.
(currently hard-coded for kk off-diagonal).

In other words, suppose we had C_ij, but not C_ji (j!=i).  We suppose C_ij=C_ji, but we cannot just copy the file for C_ij to (a file for) C_ji.  We must also modify the underlying hdf5 dataset name (latfit expects the dataset name and file name to be the same).